Schema v8: finish int32-ms migration across every timestamped field
The first v8 commit (bc7abb5) flipped PositionTrack.T and per-player
Spawns/Deaths to int32 ms — fixing the gib-respawn precision bug
locally. It deliberately deferred the rest of the timestamped fields
(ChangeI16, Interval, MatchEvent, frag/powerup event times, etc.) as
out of scope for the bug fix.

The argument for finishing the migration: schema-stabilisation. A
partial cut leaves the schema in a worse resting state than before v8
— time units are now mixed, every future field has to pick a side
and document why, and the view layer is bilingual (int32 ms vs
float64 seconds at different comparison sites). Doing both halves at
the same v8 cut is one consumer break instead of two, and the per-
break overhead (consumer migration, doc rewrites, version pin churn)
is paid once.

After this commit every timestamped field in mvd-analytics/result is
int32 milliseconds:

  - ChangeI16/ChangeI8/ChangeStr.T
  - Interval.Start/End
  - GlobalStream.MatchStart/MatchEnd
  - MatchResult.Duration/StartTime/EndTime
  - TimelineAnalysisResult.MatchStartTime/DemoOffset
  - TimelineFragEvent.Time
  - PowerupEvent.Time/EndTime/Duration
  - FragStreakEvent.Time/EndTime/Duration
  - MatchEvent.Time
  - FragEntry.Time
  - BackpackDrop.Time
  - WeaponPickup.Time/NextDeathTime/DropTime
  - ItemPhase.AvailableFrom/TakenAt/RespawnAt
  - HighResBucket.T

JSON keys are unchanged from the v8 first-cut shape; external
consumers reading these as seconds must scale by * 0.001. The
schema-version bump is the signal (CurrentSchemaVersion stays at 8 —
the first cut hasn't been released yet, so this is one expanded v8
cut, not v8 followed by v9).

Surface changes:

  - mvd-analytics/result/{streams,messages,timeline,frag,backpacks,
    weapon_pickups,match,items}.go: type flips on every time field.
  - mvd-analytics/analyzer/: added msTime(float64) int32 helper in
    timeline_streams.go (math.Round-based; well-conditioned because
    the float64-seconds view is derived once from the decoder's
    int32-ms accumulator). Producers (frag.go, messages.go,
    timeline*, items.go, weapon_pickups.go, backpacks.go) convert
    at the write site via msTime(...). Postprocess.go collapses the
    seconds→ms conversion entirely — matchStart is now read directly
    from TimelineAnalysisResult.MatchStartTime in int32 ms.
  - mvd-analytics/view/{buckets,events,streamslice,stateat,trails,
    legacy}.go: public surface (ViewBucket.T, TaggedEvent.T,
    StreamSlice opts/output, StateAt opts, LocTrails) keeps float64
    seconds — view-layer consumers (incl. the bulk of mvd-web/app.js)
    are unaffected. Internal comparisons against schema fields use
    int32 ms after a single seconds→ms conversion at function entry.
  - mvd-api/{overview,handlers}.go: Overview struct flipped to int32
    ms throughout to mirror the schema. Time-window query params
    (?from=&to=) still arrive as seconds; the handler converts to
    ms at the use site.
  - mvd-web/static/app.js: 8 read sites updated. The bulk lives in
    displayTimelineAnalysis() which projects every raw schema field
    onto timelineState in seconds — downstream panels stay
    oblivious. Direct readers (chat panel, key moments, backpack/
    weapon-pickup overlays, items panel, hub-viewer URL builders)
    convert ms→seconds at the read site. Bucket-fed panels (health/
    armor graphs, region control, map trails) are unchanged.
  - mvd-analytics/cmd/qw-analyze/main.go: Markdown report formats
    Duration/Streak/Powerup times back to seconds for display.
  - All test fixtures re-typed (int32 ms literals). golden_test.go
    sampling windows flipped to ms. Goldens regenerated.

Docs:

  - mvd-analytics/RESULT_SCHEMA.md: every field row documents int32
    ms; ChangeI16/ChangeStr/Interval shape updated; new "Schema v8:
    all times are int32 milliseconds" section enumerates the affected
    fields.
  - README.md, mvd-analytics/README.md: schema v8 paragraph rewritten
    to describe the full migration (not the partial cut).
  - mvd-analytics/INT_MS_FOLLOWUPS.md: rewritten as a record of the
    completed migration + conventions for adding new timestamped
    fields, rather than a list of deferred work.
  - FOLLOWUPS.md: deferred-work section turned into a completed-work
    note.
  - mvd-reader/MVD_FORMAT.md: notes the int-ms unit flows end-to-end.

Also fixes a stale .gitignore reference (`qwanalytics/testdata/cache/`
→ `mvd-analytics/testdata/cache/`) and adds an emacs autosave/lock
pattern.

Verification: make test green (unit + golden + diagnostic + mvd-api +
mvd-mcp); make build produces a clean WASM bundle; spot-check on the
regenerated goldens shows clean integer-ms values throughout (e.g.
"t": 6728 not "t": 6.728000000000005).

+++ mvd-analytics/analyzer/state.go
@@ -93,25 +93,27 @@ type streamBuilder struct {
 // the analyser package so tests don't have to round-trip through
 // result every time. Health/armor/loc/ammo all share int16, since
 // Quake values regularly exceed int8 range (mega-health = 200, RA = 200).
+// t is integer milliseconds (schema v8) — same unit as the result type.
 type changeI16 struct {
-	t float64
+	t int32
 	v int16
 }
 
 type changeStr struct {
-	t float64
+	t int32
 	v string
 }
 
 // intervalState tracks an open-anchor period for a boolean stream.
 // When held flips true the analyser sets anchor; on flip-false (or at
-// match end) the [anchor, t) interval is appended to closed.
+// match end) the [anchor, t) interval is appended to closed. Times
+// are integer milliseconds (schema v8).
 type intervalState struct {
-	held       bool
-	anchor     float64
-	closed     []intervalRecord
+	held   bool
+	anchor int32
+	closed []intervalRecord
 }
 
 type intervalRecord struct {
-	start, end float64
+	start, end int32
 }
